tests/fuzz_tests.py
"""
Fuzz Testing for MEV Bot
Tests edge cases and unexpected inputs
"""

import pytest
import random
from hypothesis import given, strategies as st

# Note: Requires hypothesis package
# pip install hypothesis


class TestGasCalculationFuzzing:
    """Fuzz test gas calculations"""
    
    @given(
        gas_limit=st.integers(min_value=21000, max_value=10000000),
        gas_price_gwei=st.floats(min_value=0.1, max_value=1000.0)
    )
    def test_gas_cost_calculation(self, gas_limit, gas_price_gwei):
        """Test gas cost with random inputs"""
        # Gas cost = gas_limit * gas_price
        gas_price_wei = int(gas_price_gwei * 1e9)
        gas_cost_wei = gas_limit * gas_price_wei
        
        # Should always be positive
        assert gas_cost_wei >= 0
        
        # Should not overflow (Python handles this, but test anyway)
        assert gas_cost_wei < 2**256


class TestProfitCalculationFuzzing:
    """Fuzz test profit calculations"""
    
    @given(
        amount_in=st.floats(min_value=0.001, max_value=1000000),
        price_a=st.floats(min_value=0.0001, max_value=100000),
        price_b=st.floats(min_value=0.0001, max_value=100000),
        gas_cost=st.floats(min_value=0, max_value=100)
    )
    def test_arbitrage_profit_calculation(self, amount_in, price_a, price_b, gas_cost):
        """Test arbitrage profit with random prices"""
        # Buy at price_a, sell at price_b
        amount_bought = amount_in / price_a
        amount_sold = amount_bought * price_b
        
        gross_profit = amount_sold - amount_in
        net_profit = gross_profit - gas_cost
        
        # Net profit should be calculable (not NaN)
        assert not (net_profit != net_profit)  # Check for NaN
        
        # If price_b > price_a + gas_impact, should be profitable
        if price_b > price_a * 1.01:  # At least 1% difference
            # Might be profitable (depends on gas)
            pass


class TestSlippageCalculationFuzzing:
    """Fuzz test slippage calculations"""
    
    @given(
        reserve_in=st.integers(min_value=1000, max_value=10**18),
        reserve_out=st.integers(min_value=1000, max_value=10**18),
        amount_in=st.integers(min_value=1, max_value=10**15)
    )
    def test_uniswap_v2_slippage(self, reserve_in, reserve_out, amount_in):
        """Test Uniswap V2 constant product formula"""
        # x * y = k
        # amount_out = (amount_in * 0.997 * reserve_out) / (reserve_in + amount_in * 0.997)
        
        try:
            # With 0.3% fee
            amount_in_with_fee = amount_in * 997 // 1000
            numerator = amount_in_with_fee * reserve_out
            denominator = reserve_in + amount_in_with_fee
            
            if denominator > 0:
                amount_out = numerator // denominator
                
                # Amount out should be less than reserve
                assert amount_out < reserve_out
                
                # Should not be negative
                assert amount_out >= 0
        except (ZeroDivisionError, OverflowError):
            # Division by zero or overflow - acceptable in fuzz test
            pass


class TestPriceImpactFuzzing:
    """Fuzz test price impact calculations"""
    
    @given(
        liquidity=st.floats(min_value=1000, max_value=10**9),
        trade_size=st.floats(min_value=1, max_value=10**6)
    )
    def test_price_impact_percentage(self, liquidity, trade_size):
        """Test price impact calculation"""
        if liquidity > 0:
            # Price impact ≈ trade_size / liquidity * 100
            impact_pct = (trade_size / liquidity) * 100
            
            # Impact should be between 0 and 100%
            assert 0 <= impact_pct <= 100 or impact_pct > 100
            
            # Large trades should have high impact
            if trade_size > liquidity * 0.1:
                assert impact_pct > 5  # At least 5% impact


class TestEdgeCases:
    """Test edge cases and boundary conditions"""
    
    def test_zero_liquidity(self):
        """Test behavior with zero liquidity"""
        liquidity = 0
        trade_size = 100
        
        # Should handle gracefully (not crash)
        try:
            if liquidity > 0:
                impact = trade_size / liquidity
            else:
                impact = float('inf')
            
            assert impact == float('inf') or impact > 0
        except ZeroDivisionError:
            # Acceptable - should be caught in real code
            pass
    
    def test_extremely_large_numbers(self):
        """Test with very large numbers"""
        huge_number = 10**30
        
        # Should not overflow (Python handles arbitrary precision)
        result = huge_number * 2
        assert result == 2 * 10**30
    
    def test_negative_amounts(self):
        """Test that negative amounts are rejected"""
        amount = -100
        
        # Should be rejected (in real code)
        assert amount < 0
        # In real implementation, this would raise an error
    
    def test_extremely_small_numbers(self):
        """Test with very small numbers (dust amounts)"""
        tiny_amount = 0.000000001
        
        # Should handle without precision loss
        assert tiny_amount > 0
        assert tiny_amount * 1000000000 == pytest.approx(1, rel=1e-6)


class TestConcurrentOperations:
    """Test concurrent operation scenarios"""
    
    def test_nonce_management_race_condition(self):
        """Test nonce handling in concurrent transactions"""
        # Simulate multiple transactions
        nonces = []
        current_nonce = 10
        
        for _ in range(5):
            nonce = current_nonce
            current_nonce += 1
            nonces.append(nonce)
        
        # Nonces should be sequential
        assert nonces == [10, 11, 12, 13, 14]
        
        # No duplicates
        assert len(nonces) == len(set(nonces))


class TestContractInteractionFuzzing:
    """Fuzz test contract interactions"""
    
    @given(
        amount=st.integers(min_value=1, max_value=10**18),
        slippage_bps=st.integers(min_value=1, max_value=10000)
    )
    def test_slippage_calculation(self, amount, slippage_bps):
        """Test slippage tolerance calculation"""
        # min_amount_out = amount * (10000 - slippage_bps) / 10000
        min_amount_out = (amount * (10000 - slippage_bps)) // 10000
        
        # Min amount should be less than or equal to original
        assert min_amount_out <= amount
        
        # Should not be negative
        assert min_amount_out >= 0
        
        # With 0 slippage, should equal amount
        if slippage_bps == 0:
            assert min_amount_out == amount


class TestMemoryAndPerformance:
    """Test memory usage and performance edge cases"""
    
    def test_large_list_operations(self):
        """Test with large lists (e.g., many pending transactions)"""
        # Simulate 1000 pending transactions
        pending_txs = [{'hash': f'0x{i:064x}', 'data': 'x' * 1000} for i in range(1000)]
        
        # Should handle without memory issues
        assert len(pending_txs) == 1000
        
        # Filter operations should work
        filtered = [tx for tx in pending_txs if int(tx['hash'], 16) % 2 == 0]
        assert len(filtered) == 500
    
    def test_deep_recursion(self):
        """Test recursion limits (e.g., triangular arbitrage paths)"""
        def find_paths(depth, max_depth=10):
            if depth >= max_depth:
                return [[]]
            
            paths = []
            for i in range(2):  # Binary tree
                sub_paths = find_paths(depth + 1, max_depth)
                paths.extend(sub_paths)
            
            return paths
        
        # Should complete without stack overflow
        paths = find_paths(0, 5)
        assert len(paths) > 0


class TestDataValidation:
    """Test input validation and sanitization"""
    
    @given(address=st.text())
    def test_ethereum_address_validation(self, address):
        """Test Ethereum address validation"""
        # Valid address format: 0x + 40 hex characters
        is_valid = (
            address.startswith('0x') and
            len(address) == 42 and
            all(c in '0123456789abcdefABCDEF' for c in address[2:])
        )
        
        # Should be True or False (not crash)
        assert isinstance(is_valid, bool)
    
    @given(
        token_amount=st.floats(allow_nan=True, allow_infinity=True),
        decimals=st.integers(min_value=0, max_value=18)
    )
    def test_token_amount_conversion(self, token_amount, decimals):
        """Test token amount conversions with edge cases"""
        import math
        
        # Should handle NaN and infinity
        if math.isnan(token_amount) or math.isinf(token_amount):
            # Should be detected and rejected
            assert math.isnan(token_amount) or math.isinf(token_amount)
        else:
            # Normal conversion
            if decimals >= 0:
                try:
                    wei_amount = int(token_amount * (10 ** decimals))
                    # Should be calculable
                    assert isinstance(wei_amount, int)
                except (ValueError, OverflowError):
                    # Acceptable for extreme values
                    pass


# Performance benchmark tests
class TestPerformanceBenchmarks:
    """Benchmark critical operations"""
    
    def test_multicall_performance(self):
        """Test multicall with many calls"""
        import time
        
        # Simulate 100 calls
        calls = [{'target': f'0x{i:040x}', 'data': b'\x00' * 100} for i in range(100)]
        
        start = time.time()
        
        # Process calls
        results = []
        for call in calls:
            # Simulate processing
            results.append(call['data'])
        
        elapsed = time.time() - start
        
        # Should complete in reasonable time
        assert elapsed < 1.0  # Less than 1 second
        assert len(results) == 100


# Run all tests
if __name__ == "__main__":
    pytest.main([__file__, '-v', '--tb=short'])