requirements.txt
# Core Web3 & Blockchain
web3==6.11.3
eth-account==0.10.0
eth-abi==4.2.1
eth-utils==2.3.1
hexbytes==0.3.1

# Async & Networking
aiohttp==3.9.1
asyncio==3.4.3
websockets==12.0
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Database
sqlite3  # Built-in Python

# Monitoring & Alerts
python-dotenv==1.0.0
schedule==1.2.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Security
cryptography==41.0.7

# Optional (for advanced features)
# torch==2.1.1  # Only if using deep learning
# tensorflow==2.15.0  # Only if using deep learning